requirements.txt
flask==2.2.5
numpy==1.24.3
pandas==2.2.1
scikit-learn==1.1.2
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==9.0.0
xgboost==2.1.1 
flask-cors
sklearn-pandas
seaborn==0.13.2